Flytt endepunkt inn i ForvalterController
